added different train mode components
