+++ src/lib/components/TrainView/TrainClassicSmart.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+    import { P } from 'flowbite-svelte';
+</script>
+
+<P class="text-center mt-10">Smart Cube Training Not Implemented Yet</P>
