+++ src/lib/components/TrainView/TrainDrill.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+    import { P } from 'flowbite-svelte';
+</script>
+
+<P class="text-center mt-10">Drill Mode Not Implemented Yet</P>
